Rename com.rivetz.stub -> com.rivetz.bridge
