+++ app/src/androidTest/java/com/rivetz/bridge/ApplicationTest.java
@@ -1,4 +1,4 @@
-package com.rivetz.stub;
+package com.rivetz.bridge;
 
 import android.app.Application;
 import android.test.ApplicationTestCase;
