+++ app/src/main/java/com/rivetz/bridge/KeyRecord.java
@@ -1,4 +1,4 @@
-package com.rivetz.stub;
+package com.rivetz.bridge;
 
 import android.util.Log;
