+++ app/src/main/java/com/rivetz/bridge/ServiceProviderManager.java
@@ -1,4 +1,4 @@
-package com.rivetz.stub;
+package com.rivetz.bridge;
 
 import android.content.Context;
 import android.util.Log;
