+++ app/src/main/java/com/rivetz/bridge/Utilities.java
@@ -1,4 +1,4 @@
-package com.rivetz.stub;
+package com.rivetz.bridge;
 
 import android.os.Build;
